+++ packages/account-tree-controller/src/AccountTreeController.test.ts
@@ -4,7 +4,8 @@ import type {
   MultichainAccountWalletId,
 } from '@metamask/account-api';
 import {
-  AccountWalletCategory,
+  AccountGroupType,
+  AccountWalletType,
   toAccountGroupId,
   toAccountWalletId,
   toMultichainAccountId,
@@ -27,9 +28,9 @@ import type { InternalAccount } from '@metamask/keyring-internal-api';
 import type { GetSnap as SnapControllerGetSnap } from '@metamask/snaps-controllers';
 
 import { AccountTreeController } from './AccountTreeController';
-import { AccountTreeGroup } from './AccountTreeGroup';
-import { AccountTreeWallet } from './AccountTreeWallet';
-import { EntropyRule } from './rules/entropy';
+import type { AccountGroupObject } from './group';
+import { AccountTreeGroup } from './group';
+import { BaseRule } from './rule';
 import { getAccountWalletNameFromKeyringType } from './rules/keyring';
 import {
   type AccountTreeControllerMessenger,
@@ -39,6 +40,7 @@ import {
   type AllowedActions,
   type AllowedEvents,
 } from './types';
+import { AccountTreeWallet } from './wallet';
 
 // Local mock of EMPTY_ACCOUNT to avoid circular dependency
 const EMPTY_ACCOUNT_MOCK: InternalAccount = {
@@ -397,14 +399,17 @@ describe('AccountTreeController', () => {
         MOCK_SNAP_ACCOUNT_1.options.entropy.groupIndex,
       );
       const expectedSnapWalletId = toAccountWalletId(
-        AccountWalletCategory.Snap,
+        AccountWalletType.Snap,
         MOCK_SNAP_2.id,
       );
       const expectedSnapWalletIdGroup = toAccountGroupId(
         expectedSnapWalletId,
         MOCK_SNAP_ACCOUNT_2.address,
       );
-      const expectedKeyringWalletId = `${AccountWalletCategory.Keyring}:${KeyringTypes.ledger}`;
+      const expectedKeyringWalletId = toAccountWalletId(
+        AccountWalletType.Keyring,
+        KeyringTypes.ledger,
+      );
       const expectedKeyringWalletIdGroup = toAccountGroupId(
         expectedKeyringWalletId,
         MOCK_HARDWARE_ACCOUNT_1.address,
@@ -415,18 +420,22 @@ describe('AccountTreeController', () => {
           wallets: {
             [expectedWalletId1]: {
               id: expectedWalletId1,
+              type: AccountWalletType.Entropy,
               groups: {
                 [expectedWalletId1Group]: {
                   id: expectedWalletId1Group,
+                  type: AccountGroupType.MultichainAccount,
                   accounts: [MOCK_HD_ACCOUNT_1.id],
                   metadata: {
                     name: MOCK_HD_ACCOUNT_1.metadata.name,
+                    entropy: {
+                      groupIndex: MOCK_HD_ACCOUNT_1.options.entropy.groupIndex,
+                    },
                   },
                 },
               },
               metadata: {
                 name: 'Wallet 1',
-                type: AccountWalletCategory.Entropy,
                 entropy: {
                   id: MOCK_HD_KEYRING_1.metadata.id,
                   index: 0,
@@ -435,25 +444,34 @@ describe('AccountTreeController', () => {
             },
             [expectedWalletId2]: {
               id: expectedWalletId2,
+              type: AccountWalletType.Entropy,
               groups: {
                 [expectedWalletId2Group1]: {
                   id: expectedWalletId2Group1,
+                  type: AccountGroupType.MultichainAccount,
                   accounts: [MOCK_HD_ACCOUNT_2.id],
                   metadata: {
                     name: MOCK_HD_ACCOUNT_2.metadata.name,
+                    entropy: {
+                      groupIndex: MOCK_HD_ACCOUNT_2.options.entropy.groupIndex,
+                    },
                   },
                 },
                 [expectedWalletId2Group2]: {
                   id: expectedWalletId2Group2,
+                  type: AccountGroupType.MultichainAccount,
                   accounts: [MOCK_SNAP_ACCOUNT_1.id],
                   metadata: {
                     name: MOCK_SNAP_ACCOUNT_1.metadata.name,
+                    entropy: {
+                      groupIndex:
+                        MOCK_SNAP_ACCOUNT_1.options.entropy.groupIndex,
+                    },
                   },
                 },
               },
               metadata: {
                 name: 'Wallet 2',
-                type: AccountWalletCategory.Entropy,
                 entropy: {
                   id: MOCK_HD_KEYRING_2.metadata.id,
                   index: 1,
@@ -462,9 +480,11 @@ describe('AccountTreeController', () => {
             },
             [expectedSnapWalletId]: {
               id: expectedSnapWalletId,
+              type: AccountWalletType.Snap,
               groups: {
                 [expectedSnapWalletIdGroup]: {
                   id: expectedSnapWalletIdGroup,
+                  type: AccountGroupType.SingleAccount,
                   accounts: [MOCK_SNAP_ACCOUNT_2.id],
                   metadata: {
                     name: MOCK_SNAP_ACCOUNT_2.metadata.name,
@@ -473,7 +493,6 @@ describe('AccountTreeController', () => {
               },
               metadata: {
                 name: MOCK_SNAP_2.manifest.proposedName,
-                type: AccountWalletCategory.Snap,
                 snap: {
                   id: MOCK_SNAP_2.id,
                 },
@@ -481,9 +500,11 @@ describe('AccountTreeController', () => {
             },
             [expectedKeyringWalletId]: {
               id: expectedKeyringWalletId,
+              type: AccountWalletType.Keyring,
               groups: {
                 [expectedKeyringWalletIdGroup]: {
                   id: expectedKeyringWalletIdGroup,
+                  type: AccountGroupType.SingleAccount,
                   accounts: [MOCK_HARDWARE_ACCOUNT_1.id],
                   metadata: {
                     name: MOCK_HARDWARE_ACCOUNT_1.metadata.name,
@@ -494,7 +515,6 @@ describe('AccountTreeController', () => {
                 name: getAccountWalletNameFromKeyringType(
                   MOCK_HARDWARE_ACCOUNT_1.metadata.keyring.type as KeyringTypes,
                 ),
-                type: AccountWalletCategory.Keyring,
                 keyring: {
                   type: KeyringTypes.ledger,
                 },
@@ -567,7 +587,7 @@ describe('AccountTreeController', () => {
       // Since no entropy sources will be found, it will be categorized as a
       // "Keyring" wallet
       const wallet1Id = toAccountWalletId(
-        AccountWalletCategory.Snap,
+        AccountWalletType.Snap,
         MOCK_SNAP_1.id,
       );
 
@@ -598,11 +618,11 @@ describe('AccountTreeController', () => {
       // Since no entropy sources will be found, it will be categorized as a
       // "Keyring" wallet
       const wallet1Id = toAccountWalletId(
-        AccountWalletCategory.Keyring,
+        AccountWalletType.Keyring,
         mockHdAccount1.metadata.keyring.type,
       );
       const wallet2Id = toAccountWalletId(
-        AccountWalletCategory.Keyring,
+        AccountWalletType.Keyring,
         mockHdAccount1.metadata.keyring.type,
       );
 
@@ -664,16 +684,22 @@ describe('AccountTreeController', () => {
           wallets: {
             [walletId1]: {
               id: walletId1,
+              type: AccountWalletType.Entropy,
               groups: {
                 [walletId1Group]: {
                   id: walletId1Group,
-                  metadata: { name: mockHdAccount1.metadata.name },
+                  type: AccountGroupType.MultichainAccount,
+                  metadata: {
+                    name: mockHdAccount1.metadata.name,
+                    entropy: {
+                      groupIndex: mockHdAccount1.options.entropy.groupIndex,
+                    },
+                  },
                   accounts: [mockHdAccount2.id], // HD account 1 got removed.
                 },
               },
               metadata: {
                 name: 'Wallet 1',
-                type: AccountWalletCategory.Entropy,
                 entropy: {
                   id: MOCK_HD_KEYRING_1.metadata.id,
                   index: 0,
@@ -736,16 +762,22 @@ describe('AccountTreeController', () => {
           wallets: {
             [walletId1]: {
               id: walletId1,
+              type: AccountWalletType.Entropy,
               groups: {
                 [walletId1Group]: {
                   id: walletId1Group,
-                  metadata: { name: mockHdAccount1.metadata.name },
+                  type: AccountGroupType.MultichainAccount,
+                  metadata: {
+                    name: mockHdAccount1.metadata.name,
+                    entropy: {
+                      groupIndex: mockHdAccount1.options.entropy.groupIndex,
+                    },
+                  },
                   accounts: [mockHdAccount1.id, mockHdAccount2.id], // HD account 2 got added.
                 },
               },
               metadata: {
                 name: 'Wallet 1',
-                type: AccountWalletCategory.Entropy,
                 entropy: {
                   id: MOCK_HD_KEYRING_1.metadata.id,
                   index: 0,
@@ -813,16 +845,22 @@ describe('AccountTreeController', () => {
           wallets: {
             [walletId1]: {
               id: walletId1,
+              type: AccountWalletType.Entropy,
               groups: {
                 [walletId1Group]: {
                   id: walletId1Group,
-                  metadata: { name: mockHdAccount1.metadata.name },
+                  type: AccountGroupType.MultichainAccount,
+                  metadata: {
+                    name: mockHdAccount1.metadata.name,
+                    entropy: {
+                      groupIndex: mockHdAccount1.options.entropy.groupIndex,
+                    },
+                  },
                   accounts: [mockHdAccount1.id],
                 },
               },
               metadata: {
                 name: 'Wallet 1',
-                type: AccountWalletCategory.Entropy,
                 entropy: {
                   id: MOCK_HD_KEYRING_1.metadata.id,
                   index: 0,
@@ -832,16 +870,22 @@ describe('AccountTreeController', () => {
             [walletId2]: {
               // New wallet automatically added.
               id: walletId2,
+              type: AccountWalletType.Entropy,
               groups: {
                 [walletId2Group]: {
                   id: walletId2Group,
-                  metadata: { name: mockHdAccount2.metadata.name },
+                  type: AccountGroupType.MultichainAccount,
+                  metadata: {
+                    name: mockHdAccount2.metadata.name,
+                    entropy: {
+                      groupIndex: mockHdAccount2.options.entropy.groupIndex,
+                    },
+                  },
                   accounts: [mockHdAccount2.id],
                 },
               },
               metadata: {
                 name: 'Wallet 2',
-                type: AccountWalletCategory.Entropy,
                 entropy: {
                   id: MOCK_HD_KEYRING_2.metadata.id,
                   index: 1,
@@ -864,7 +908,7 @@ describe('AccountTreeController', () => {
       controller.init();
 
       const walletId = toAccountWalletId(
-        AccountWalletCategory.Entropy,
+        AccountWalletType.Entropy,
         MOCK_HD_KEYRING_1.metadata.id,
       );
       const wallet = controller.getAccountWallet(walletId);
@@ -912,7 +956,7 @@ describe('AccountTreeController', () => {
       const wallet = wallets[0];
       expect(wallet.id).toBeDefined();
       expect(wallet.name).toBeDefined();
-      expect(wallet.category).toBeDefined();
+      expect(wallet.type).toBeDefined();
 
       const groups = wallet.getAccountGroups();
       expect(groups).toHaveLength(1);
@@ -984,6 +1028,7 @@ describe('AccountTreeController', () => {
       expect(group.id).toBeDefined();
       expect(group.wallet).toBeDefined();
       expect(group.name).toBeDefined();
+      expect(group.type).toBeDefined();
 
       const accounts = group.getAccounts();
       const accountIds = group.getAccountIds();
@@ -1020,14 +1065,11 @@ describe('AccountTreeController', () => {
       const wallet = new AccountTreeWallet({
         messenger,
         wallet: {
-          id: toAccountWalletId(
-            AccountWalletCategory.Keyring,
-            KeyringTypes.simple,
-          ),
+          id: toAccountWalletId(AccountWalletType.Keyring, KeyringTypes.simple),
+          type: AccountWalletType.Keyring,
           groups: {},
           metadata: {
             name: '',
-            type: AccountWalletCategory.Keyring,
             keyring: {
               type: KeyringTypes.simple,
             },
@@ -1039,6 +1081,7 @@ describe('AccountTreeController', () => {
         wallet,
         group: {
           id: toAccountGroupId(wallet.id, 'bad'),
+          type: AccountGroupType.SingleAccount,
           accounts: [account.id],
           metadata: {
             name: '',
@@ -1053,55 +1096,17 @@ describe('AccountTreeController', () => {
       expect(group.getOnlyAccount()).toBe(account);
     });
 
-    it('throws if the group has no account', () => {
-      const messenger = getAccountTreeControllerMessenger();
-
-      const wallet = new AccountTreeWallet({
-        messenger,
-        wallet: {
-          id: toAccountWalletId(
-            AccountWalletCategory.Keyring,
-            KeyringTypes.simple,
-          ),
-          groups: {},
-          metadata: {
-            name: '',
-            type: AccountWalletCategory.Keyring,
-            keyring: {
-              type: KeyringTypes.simple,
-            },
-          },
-        },
-      });
-      const group = new AccountTreeGroup({
-        messenger,
-        wallet,
-        group: {
-          id: toAccountGroupId(wallet.id, 'bad'),
-          accounts: [],
-          metadata: {
-            name: '',
-          },
-        },
-      });
-
-      expect(() => group.getOnlyAccount()).toThrow('Group contains no account');
-    });
-
     it('throws if the group has more than 1 account when calling getOnlyAccount', () => {
       const messenger = getAccountTreeControllerMessenger();
 
       const wallet = new AccountTreeWallet({
         messenger,
         wallet: {
-          id: toAccountWalletId(
-            AccountWalletCategory.Keyring,
-            KeyringTypes.simple,
-          ),
+          id: toAccountWalletId(AccountWalletType.Keyring, KeyringTypes.simple),
+          type: AccountWalletType.Keyring,
           groups: {},
           metadata: {
             name: '',
-            type: AccountWalletCategory.Keyring,
             keyring: {
               type: KeyringTypes.simple,
             },
@@ -1113,7 +1118,11 @@ describe('AccountTreeController', () => {
         wallet,
         group: {
           id: toAccountGroupId(wallet.id, 'bad'),
-          accounts: [MOCK_HD_ACCOUNT_1.id, MOCK_HD_ACCOUNT_2.id],
+          type: AccountGroupType.SingleAccount,
+          // Testing an error case here, so we have to cast.
+          accounts: [MOCK_HD_ACCOUNT_1.id, MOCK_HD_ACCOUNT_2.id] as unknown as [
+            InternalAccount['id'],
+          ],
           metadata: {
             name: '',
           },
@@ -1535,74 +1544,34 @@ describe('AccountTreeController', () => {
       expect(controller.getSelectedAccountGroup()).not.toBe('');
     });
   });
-});
 
-describe('AccountTreeRule', () => {
-  const account = MOCK_HD_ACCOUNT_1;
-  const group = {
-    id: toAccountGroupId(
-      toAccountWalletId(AccountWalletCategory.Entropy, 'test'),
-      'test',
-    ),
-    accounts: [account.id],
-    metadata: {
-      name: '',
-    },
-  };
+  describe('BaseRule', () => {
+    it('fallbacks to emptry group name if we cannot get its account', () => {
+      const rootMessenger = getRootMessenger();
+      const messenger = getAccountTreeControllerMessenger(rootMessenger);
+      const rule = new BaseRule(messenger);
 
-  const setupRule = () => {
-    const messenger = getRootMessenger();
+      rootMessenger.registerActionHandler(
+        'AccountsController:getAccount',
+        () => undefined,
+      );
 
-    return {
-      rule: new EntropyRule(getAccountTreeControllerMessenger(messenger)),
-      messenger,
-    };
-  };
+      const group: AccountGroupObject = {
+        id: toMultichainAccountId(
+          toMultichainAccountWalletId('test'),
+          MOCK_HD_ACCOUNT_1.options.entropy.groupIndex,
+        ),
+        type: AccountGroupType.MultichainAccount,
+        accounts: [MOCK_HD_ACCOUNT_1.id],
+        metadata: {
+          name: MOCK_HD_ACCOUNT_1.metadata.name,
+          entropy: {
+            groupIndex: MOCK_HD_ACCOUNT_1.options.entropy.groupIndex,
+          },
+        },
+      };
 
-  it('gets accounts from a group', () => {
-    const { messenger, rule } = setupRule();
-
-    messenger.registerActionHandler(
-      'AccountsController:getAccount',
-      () => MOCK_HD_ACCOUNT_1,
-    );
-
-    expect(rule.getOnlyAccountFrom(group)).toStrictEqual(MOCK_HD_ACCOUNT_1);
-    expect(rule.getAccountsFrom(group)).toStrictEqual([MOCK_HD_ACCOUNT_1]);
-  });
-
-  it('throws if it cannot get account', () => {
-    const { messenger, rule } = setupRule();
-
-    messenger.registerActionHandler(
-      'AccountsController:getAccount',
-      () => undefined,
-    );
-
-    expect(() => rule.getOnlyAccountFrom(group)).toThrow(
-      `Unable to get account with ID: "${account.id}"`,
-    );
-  });
-
-  it('throws if there is not enough accounts', () => {
-    const { rule } = setupRule();
-
-    expect(() =>
-      rule.getOnlyAccountFrom({
-        ...group,
-        accounts: [],
-      }),
-    ).toThrow('Group contains no account');
-  });
-
-  it('throws if there is too many accounts', () => {
-    const { rule } = setupRule();
-
-    expect(() =>
-      rule.getOnlyAccountFrom({
-        ...group,
-        accounts: [MOCK_HD_ACCOUNT_1.id, MOCK_HD_ACCOUNT_2.id],
-      }),
-    ).toThrow('Group contains more than 1 account');
+      expect(rule.getDefaultAccountGroupName(group)).toBe('');
+    });
   });
 });